+++ final_counts.txt
@@ -1,0 +1,3 @@
+﻿src\features\data-retention\__tests__\archive-search-queries.admissions.test.ts: 230
+src\features\data-retention\__tests__\archive-search-queries.audit.test.ts: 219
+Checked 429 files.
