+++ final_counts_2.txt
@@ -1,0 +1,3 @@
+﻿src\features\data-retention\__tests__\archive-search-queries.admissions.test.ts: 217
+src\features\data-retention\__tests__\archive-search-queries.audit.test.ts: 205
+Checked 430 files.
