+++ lint_output.txt
@@ -1,0 +1,4 @@
+﻿
+> ewtcs@0.1.0 lint
+> next lint
+
